+++ Slim/Routing/Dispatcher.php
@@ -9,7 +9,7 @@
 
 declare(strict_types=1);
 
-namespace Slim;
+namespace Slim\Routing;
 
 use FastRoute\Dispatcher\GroupCountBased;
